nested ophys interfaces according to format
