+++ nwb_conversion_tools/datainterfaces/ophys/caiman/caimandatainterface.py
@@ -1,0 +1,15 @@
+from roiextractors import (
+    CnmfeSegmentationExtractor,
+    ExtractSegmentationExtractor,
+    CaimanSegmentationExtractor,
+    Suite2pSegmentationExtractor,
+    SimaSegmentationExtractor
+)
+
+from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
+
+
+class CaimanSegmentationInterface(BaseSegmentationExtractorInterface):
+    """Data interface for CaimanSegmentationExtractor"""
+
+    SegX = CaimanSegmentationExtractor
