+++ nwb_conversion_tools/datainterfaces/ophys/cnmfe/cnmfedatainterface.py
@@ -1,0 +1,15 @@
+from roiextractors import (
+    CnmfeSegmentationExtractor,
+    ExtractSegmentationExtractor,
+    CaimanSegmentationExtractor,
+    Suite2pSegmentationExtractor,
+    SimaSegmentationExtractor
+)
+
+from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
+
+
+class CnmfeSegmentationInterface(BaseSegmentationExtractorInterface):
+    """Data interface for CnmfeRecordingInterface"""
+
+    SegX = CnmfeSegmentationExtractor
