+++ nwb_conversion_tools/datainterfaces/ophys/extract/extractdatainterface.py
@@ -1,0 +1,14 @@
+from roiextractors import (
+    CnmfeSegmentationExtractor,
+    ExtractSegmentationExtractor,
+    CaimanSegmentationExtractor,
+    Suite2pSegmentationExtractor,
+    SimaSegmentationExtractor
+)
+
+from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
+
+class ExtractSegmentationInterface(BaseSegmentationExtractorInterface):
+    """Data interface for ExtractSegmentationExtractor"""
+
+    SegX = ExtractSegmentationExtractor
