--- nwb_conversion_tools/datainterfaces/ophys/roi/roiextractordatainterface.py
@@ -1,39 +1,0 @@
-from roiextractors import (
-    CnmfeSegmentationExtractor,
-    ExtractSegmentationExtractor,
-    CaimanSegmentationExtractor,
-    Suite2pSegmentationExtractor,
-    SimaSegmentationExtractor
-)
-
-from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
-
-
-class CnmfeSegmentationInterface(BaseSegmentationExtractorInterface):
-    """Data interface for CnmfeRecordingInterface"""
-
-    SegX = CnmfeSegmentationExtractor
-
-
-class ExtractSegmentationInterface(BaseSegmentationExtractorInterface):
-    """Data interface for ExtractSegmentationExtractor"""
-
-    SegX = ExtractSegmentationExtractor
-
-
-class CaimanSegmentationInterface(BaseSegmentationExtractorInterface):
-    """Data interface for CaimanSegmentationExtractor"""
-
-    SegX = CaimanSegmentationExtractor
-
-
-class Suite2pSegmentationInterface(BaseSegmentationExtractorInterface):
-    """Data interface for Suite2pSegmentationExtractor"""
-
-    SegX = Suite2pSegmentationExtractor
-
-
-class SimaSegmentationInterface(BaseSegmentationExtractorInterface):
-    """Data interface for SimaSegmentationExtractor"""
-
-    SegX = SimaSegmentationExtractor
