+++ nwb_conversion_tools/datainterfaces/ophys/sbx/sbxdatainterface.py
@@ -1,0 +1,9 @@
+from roiextractors import SbxImagingExtractor, Hdf5ImagingExtractor, TiffImagingExtractor
+
+from ..baseimagingextractorinterface import BaseImagingExtractorInterface
+
+
+class SbxImagingInterface(BaseImagingExtractorInterface):
+    """Data Interface for SbxImagingExtractor"""
+
+    IX = SbxImagingExtractor
