+++ nwb_conversion_tools/datainterfaces/ophys/sima/simadatainterface.py
@@ -1,0 +1,15 @@
+from roiextractors import (
+    CnmfeSegmentationExtractor,
+    ExtractSegmentationExtractor,
+    CaimanSegmentationExtractor,
+    Suite2pSegmentationExtractor,
+    SimaSegmentationExtractor
+)
+
+from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
+
+
+class SimaSegmentationInterface(BaseSegmentationExtractorInterface):
+    """Data interface for SimaSegmentationExtractor"""
+
+    SegX = SimaSegmentationExtractor
