+++ nwb_conversion_tools/datainterfaces/ophys/tiff/tiffdatainterface.py
@@ -1,0 +1,9 @@
+from roiextractors import SbxImagingExtractor, Hdf5ImagingExtractor, TiffImagingExtractor
+
+from ..baseimagingextractorinterface import BaseImagingExtractorInterface
+
+
+class TiffImagingInterface(BaseImagingExtractorInterface):
+    """Data Interface for TIffImagingExtractor"""
+
+    IX = TiffImagingExtractor
